src/backend/bedrock_alerts/requirements.txt
boto3
python-dotenv
jsonlines
pytest-mock
moto
